% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgsEnrichment.R
\name{cpGsEnrichment}
\alias{cpGsEnrichment}
\title{Calculate probes enrichment}
\usage{
cpGsEnrichment(
  fg.probes,
  bg.probes,
  fg.label = "foreground",
  bg.label = "background",
  tab.filename,
  arrayType = c("450k", "EPIC"),
  save.plot = TRUE,
  plot.filename = "barplot.pdf",
  enrichment.type = c("island", "gene")
)
}
\arguments{
\item{fg.probes}{foreground probes (i.e. hypo methylated pobes)}

\item{bg.probes}{backgound probes (i.e. all comethylated probes)}

\item{fg.label}{Foreground label}

\item{bg.label}{Background label}

\item{tab.filename}{Table file name (csv file). Default no file will be output.}

\item{arrayType}{Type of array, 450k or EPIC}

\item{save.plot}{create a bar plot of frequncy for foreground and background ? Default TRUE.}

\item{plot.filename}{filename of the barplot}

\item{type}{of enrichment to caculate: relation to island ("island") or
UCSC RefGene Group mapped to the probe ("gene")}
}
\value{
A list with a ggplot2 barplo, a table and the list of contigency tables
}
\description{
This function calculates the enrichment of cpgs compared to a background
}
\examples{
data(betasChr22_df)
result.list <- cpGsEnrichment(fg.probes = rownames(betasChr22_df)[1:100],
                              bg.probes = rownames(betasChr22_df)[-c(1:100)],
                              enrichment.type = "island",
                              arrayType = "450k",
                              save.plot = FALSE)
}
