% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgsEnrichment.R
\name{cpGsGenomicFeatures}
\alias{cpGsGenomicFeatures}
\title{Plot percentage of CpGs for diff. genomic features}
\usage{
cpGsGenomicFeatures(
  probes.list,
  arrayType = c("450k", "EPIC"),
  save.plot = TRUE,
  bar.colors,
  plot.width = 7,
  plot.height = 7,
  plot.title = NULL,
  annotation.gr = NULL,
  enrichment.type = c("island_gene", "customized"),
  plot.filename = "barplot.pdf"
)
}
\arguments{
\item{arrayType}{Type of array, 450k or EPIC}

\item{save.plot}{create a bar plot of frequncy ? Default TRUE.}

\item{bar.colors}{Colors of the bars of each group}

\item{plot.width}{Plot width}

\item{plot.title}{A string title to the plot}

\item{plot.filename}{filename of the barplot}

\item{probes}{list of named probes}

\item{plot.hight}{Plot height}
}
\value{
A list with a ggplot2 barplo, a table and the list of contigency tables
}
\description{
creates barplot for percentage of dff. methylated CpGs
for enhancer, intergenic, intron, island, and other.
}
\examples{
data(betasChr22_df)
probes.list <- list("Hypomethylated CpGs" = rownames(betasChr22_df)[1:100],
                    "Hypermethylated CpgS" = rownames(betasChr22_df)[100:200])
result.list <- cpGsGenomicFeatures(probes.list = probes.list,
                               arrayType = "450k",
                                    bar.colors = c("#1F77B4", "#FF7F0D"),
                                   save.plot = FALSE)
}
