% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCpGsInRegion.R
\name{GetCpGsInAllRegion}
\alias{GetCpGsInAllRegion}
\title{Extract probe IDs for CpGs located in a genomic region}
\usage{
GetCpGsInAllRegion(
  regionName_char,
  arrayType = c("450k", "EPIC"),
  nCores_int = 1L,
  ...
)
}
\arguments{
\item{regionName_char}{character string with location information for one region in
this format: "chrxx:xxxxxx-xxxxxx"}

\item{arrayType}{Type of array, 450k or EPIC}

\item{nCores_int}{Number of computing cores to be used when executing code
in parallel. Defaults to 1 (serial computing).}

\item{...}{Dots for additional arguments passed to the cluster constructor.
See \code{\link{CreateParallelWorkers}} for more information.}
}
\value{
vector of CpG probe IDs mapped to the genomic region
}
\description{
Extract probe IDs for CpGs located in a genomic region
}
\examples{
   GetCpGsInAllRegion(
     regionName_char = c("chr14:74815131-74815316",
     "chr22:18267969-18268249"),
     arrayType = "450k"
   )
}
