% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util5_getProbesCluster.R
\name{getPredefinedCluster}
\alias{getPredefinedCluster}
\title{Get pre-annotated clusters of probes}
\usage{
getPredefinedCluster(
  arrayType = c("450k", "EPIC"),
  clusterType = c("regions", "gene")
)
}
\arguments{
\item{arrayType}{Type of array, 450k or EPIC}

\item{clusterType}{Type of cluster of probes: "gene" Cpgs or "regions" Cpgs
clusters.}
}
\value{
list of clustered probes
}
\description{
Get pre-annotated clusters of probes
}
\examples{
 cluster <- getPredefinedCluster ("450k")
}
