% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minPairwiseCor.R
\name{minPairwiseCor}
\alias{minPairwiseCor}
\title{Calculate min Pairwise correlation for subregions probes}
\usage{
minPairwiseCor(betaCluster_mat, betaToM = TRUE, minPairwiseCorr = 0.2,
  probes.list, method = c("pearson", "spearman"))
}
\arguments{
\item{betaCluster_mat}{matrix of beta values, with rownames = sample ids,
column names = CpG ids. Note that the CpGs need to be ordered by their genomic positions,
this can be accomplished by the \code{OrderCpGbyLocation} function.}

\item{betaToM}{indicates if converting to mvalues before computing correlations}

\item{minPairwiseCorr}{thershold for min pairwise correlation between any cpg the
rest of the CpGs}

\item{probes.list}{list of regions with cpgs}

\item{method}{correlation method, can be pearson or spearman}
}
\value{
A list with the list of probes passing the minPairwiseCorr and a dataframe
with the minPairwiseCorr for each subregion if subregions.annot was provided
}
\description{
Calculate min Pairwise correlation for subregions probes
}
\examples{
data(betaMatrix_ex1)
probes <- colnames(betaMatrix_ex1)
minPairwiseCor(betaCluster_mat = betaMatrix_ex1,
               betaToM = FALSE,
               minPairwiseCorr = 0.1,
               probes.list = list("chr22:18531243-18531447" = probes),
               method = "pearson")
minPairwiseCor(betaCluster_mat = betaMatrix_ex1,
               betaToM = FALSE,
               minPairwiseCorr = 0.2,
               probes.list = list("chr22:18531243-18531447" = probes),
               method = "pearson")
}
